Allow time_recorder to support re-entrancy.
The @time_recorder decorator currently maintains the start time of an
operation as a class attribute; this means that if any operations using
time_recorder are nested -- which also includes recursion -- the
recorded timing information will be incorrect.
This commit modifies @time_recorder to maintain a stack of timestamps
instead of a single timestamp to support re-entrancy. The list/stack
is also made an instance attribute instead of a class attribute,
to further avoid the possibility of having one decorated operation
overwrite another operation's information. It might, however, be
useful to undo the latter change at some point if there's interest in
tracking how the operations are nested, a la OpenTelemetry tracing.